pyproject.toml
[project]
name = "autopi-api"
version = "0.1.0"
description = "A custom reimplementation of autopi.io API"
authors = [
    {name = "monkeyclass",email = "30174727+monkeyclass@users.noreply.github.com"}
]
license = {text = "Apache"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "pydantic (>=2.10.5,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
